refactor: change some code according to the comments
