--- components/forms/fromError/index.ts
@@ -1,1 +1,0 @@
-export { FormError } from './FormError';
