+++ containers/forms/Login.tsx
@@ -8,8 +8,8 @@ import Link from 'next/link';
 import { useRouter } from 'next/navigation';
 import { useForm } from 'react-hook-form';
 import { yupResolver } from '@hookform/resolvers/yup';
-import { schema } from '@/components/forms/formValidation';
-import { Button } from '../ui/button';
+import { schema } from '@/containers/forms/formValidation';
+import { Button } from '@/components/ui/button';
 import { FormError } from './fromError';
 import { IFormData } from './types';
 import styles from './form.module.css';
