+++ containers/forms/fromError/index.ts
@@ -1,0 +1,1 @@
+export { FormError } from './formError';
